vktrace: Move vkDisplayXcb/Wayland to loadable libraries
In order to simplify loading of the Xcb and Wayland symbols at runtime,
the display classes have been moved to separate libraries that can be
statically linked to the windowing system libraries, this limits
the number of dynamic symbols needed.
