+++ vktrace/vktrace_replay/vkreplay_vkreplay.cpp
@@ -52,7 +52,13 @@ vkReplay::vkReplay(vkreplayer_settings *pReplaySettings, vktrace_trace_file_head
     m_pDSDump = NULL;
     m_pCBDump = NULL;
     m_display = display;
-    m_displayServer = pReplaySettings->displayServer;
+
+    if (strcasecmp(pReplaySettings->displayServer, "xcb") == 0) {
+        m_displayServer = VK_DISPLAY_XCB;
+    } else if (strcasecmp(pReplaySettings->displayServer, "wayland") == 0) {
+        m_displayServer = VK_DISPLAY_WAYLAND;
+    }
+
     //    m_pVktraceSnapshotPrint = NULL;
     m_objMapper.m_adjustForGPU = false;
 
@@ -3527,7 +3533,7 @@ VkResult vkReplay::manually_replay_vkCreateXlibSurfaceKHR(packet_vkCreateXlibSur
 // TODO
 #elif defined(PLATFORM_LINUX)
 /*if defined(VK_USE_PLATFORM_XLIB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXlib::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XLIB) {
         VkIcdSurfaceXlib *pSurf = (VkIcdSurfaceXlib *)m_display->get_surface();
         VkXlibSurfaceCreateInfoKHR createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_XLIB_SURFACE_CREATE_INFO_KHR;
@@ -3539,7 +3545,7 @@ VkResult vkReplay::manually_replay_vkCreateXlibSurfaceKHR(packet_vkCreateXlibSur
     }
 #endif*/
 #if defined(VK_USE_PLATFORM_XCB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         VkIcdSurfaceXcb *pSurf = (VkIcdSurfaceXcb *)m_display->get_surface();
         VkXcbSurfaceCreateInfoKHR createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR;
@@ -3551,7 +3557,7 @@ VkResult vkReplay::manually_replay_vkCreateXlibSurfaceKHR(packet_vkCreateXlibSur
     }
 #endif
 #if defined(VK_USE_PLATFORM_WAYLAND_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayWayland::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_WAYLAND) {
         VkIcdSurfaceWayland *pSurf = (VkIcdSurfaceWayland *)m_display->get_surface();
         VkWaylandSurfaceCreateInfoKHR createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR;
@@ -3624,7 +3630,7 @@ VkResult vkReplay::manually_replay_vkCreateWaylandSurfaceKHR(packet_vkCreateWayl
 // TODO
 #elif defined(PLATFORM_LINUX)
 #if defined(VK_USE_PLATFORM_WAYLAND_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayWayland::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_WAYLAND) {
         VkIcdSurfaceWayland *pSurf = (VkIcdSurfaceWayland *)m_display->get_surface();
         VkWaylandSurfaceCreateInfoKHR createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR;
@@ -3636,7 +3642,7 @@ VkResult vkReplay::manually_replay_vkCreateWaylandSurfaceKHR(packet_vkCreateWayl
     }
 #endif
 #if defined(VK_USE_PLATFORM_XCB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         VkIcdSurfaceXcb *pSurf = (VkIcdSurfaceXcb *)m_display->get_surface();
         VkXcbSurfaceCreateInfoKHR createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR;
@@ -3648,7 +3654,7 @@ VkResult vkReplay::manually_replay_vkCreateWaylandSurfaceKHR(packet_vkCreateWayl
     }
 #endif
 /*#if defined(VK_USE_PLATFORM_XLIB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         VkIcdSurfaceXlib *pSurf = (VkIcdSurfaceXlib *)m_display->get_surface();
         VkXlibSurfaceCreateInfoKHR createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_XLIB_SURFACE_CREATE_INFO_KHR;
@@ -3671,8 +3677,7 @@ VkResult vkReplay::manually_replay_vkCreateWaylandSurfaceKHR(packet_vkCreateWayl
     createInfo.hinstance = pSurf->hinstance;
     createInfo.hwnd = pSurf->hwnd;
     replayResult = m_vkFuncs.CreateWin32SurfaceKHR(remappedinstance, &createInfo, pPacket->pAllocator, &local_pSurface);
-#else
-    vktrace_LogError("manually_replay_vkCreateWaylandSurfaceKHR not implemented on this playback platform");
+#else vktrace_LogError("manually_replay_vkCreateWaylandSurfaceKHR not implemented on this playback platform");
     replayResult = VK_ERROR_FEATURE_NOT_PRESENT;
 #endif
     if (replayResult == VK_SUCCESS) {
@@ -3919,7 +3924,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceXcbPresentationSupportKHR(
     return VK_TRUE;
 #elif defined(PLATFORM_LINUX)
 #if defined(VK_USE_PLATFORM_XCB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         vkDisplayXcb *pDisp = (vkDisplayXcb *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceXcbPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                      pDisp->get_connection_handle(),
@@ -3927,7 +3932,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceXcbPresentationSupportKHR(
     }
 #endif
 /*#if defined(VK_USE_PLATFORM_XLIB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXlib::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XLIB) {
         VkIcdSurfaceXlib *pSurf = (VkIcdSurfaceXlib *)m_display->get_surface();
         return (m_vkFuncs.GetPhysicalDeviceXlibPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                              pSurf->dpy,
@@ -3935,7 +3940,7 @@ m_display->get_screen_handle()->root_visual));
     }
 #endif*/
 #if defined(VK_USE_PLATFORM_WAYLAND_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayWayland::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_WAYLAND) {
         vkDisplayWayland *pDisp = (vkDisplayWayland *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceWaylandPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                          pDisp->get_display_handle()));
@@ -3987,7 +3992,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceXlibPresentationSupportKHR
     return VK_TRUE;
 #elif defined(PLATFORM_LINUX)
 #if defined(VK_USE_PLATFORM_XCB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         vkDisplayXcb *pDisp = (vkDisplayXcb *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceXcbPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                      pDisp->get_connection_handle(),
@@ -3995,7 +4000,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceXlibPresentationSupportKHR
     }
 #endif
 /*#if defined(VK_USE_PLATFORM_XLIB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXlib::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XLIB) {
         VkIcdSurfaceXlib *pSurf = (VkIcdSurfaceXlib *)m_display->get_surface();
         return (m_vkFuncs.GetPhysicalDeviceXlibPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                              pSurf->dpy,
@@ -4003,7 +4008,7 @@ m_display->get_screen_handle()->root_visual));
     }
 #endif*/
 #if defined(VK_USE_PLATFORM_WAYLAND_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayWayland::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_WAYLAND) {
         vkDisplayWayland *pDisp = (vkDisplayWayland *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceWaylandPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                          pDisp->get_display_handle()));
@@ -4055,7 +4060,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceWaylandPresentationSupport
     return VK_TRUE;
 #elif defined(PLATFORM_LINUX)
 #if defined(VK_USE_PLATFORM_XCB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         vkDisplayXcb *pDisp = (vkDisplayXcb *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceXcbPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                      pDisp->get_connection_handle(),
@@ -4063,7 +4068,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceWaylandPresentationSupport
     }
 #endif
 /*#if defined(VK_USE_PLATFORM_XLIB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXlib::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XLIB) {
         VkIcdSurfaceXlib *pSurf = (VkIcdSurfaceXlib *)m_display->get_surface();
         return (m_vkFuncs.GetPhysicalDeviceXlibPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                              pSurf->dpy,
@@ -4071,7 +4076,7 @@ m_display->get_screen_handle()->root_visual));
     }
 #endif*/
 #if VK_USE_PLATFORM_WAYLAND_KHR
-    if (strcasecmp(m_displayServer, vkDisplayWayland::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_WAYLAND) {
         vkDisplayWayland *pDisp = (vkDisplayWayland *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceWaylandPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                          pDisp->get_display_handle()));
@@ -4123,7 +4128,7 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceWin32PresentationSupportKH
     return VK_TRUE;
 #elif defined(PLATFORM_LINUX)
 #if defined(VK_USE_PLATFORM_XCB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXcb::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XCB) {
         vkDisplayXcb *pDisp = (vkDisplayXcb *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceXcbPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                      pDisp->get_connection_handle(),
@@ -4131,14 +4136,14 @@ VkBool32 vkReplay::manually_replay_vkGetPhysicalDeviceWin32PresentationSupportKH
     }
 #endif
 /*#if defined(VK_USE_PLATFORM_XLIB_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayXlib::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_XLIB) {
         VkIcdSurfaceXlib *pSurf = (VkIcdSurfaceXlib *)m_display->get_surface();
         return (m_vkFuncs.GetPhysicalDeviceXlibPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                          pSurf->dpy, m_display->get_screen_handle()->root_visual));
     }
 #endif*/
 #if defined(VK_USE_PLATFORM_WAYLAND_KHR)
-    if (strcasecmp(m_displayServer, vkDisplayWayland::NAME) == 0) {
+    if (m_displayServer == VK_DISPLAY_WAYLAND) {
         vkDisplayWayland *pDisp = (vkDisplayWayland *)m_display;
         return (m_vkFuncs.GetPhysicalDeviceWaylandPresentationSupportKHR(remappedphysicalDevice, pPacket->queueFamilyIndex,
                                                                          pDisp->get_display_handle()));
